Extract galaxy to seperate file, introduce shorthand imports
