+++ src/astronomicals/system.rs
@@ -1,9 +1,7 @@
-use astronomicals::{hash, planet::Planet, star::Star};
+use super::*;
 use entities::Faction;
-use game::Updatable;
 use std::{fmt,
           hash::{Hash, Hasher}};
-use utils::Point;
 
 #[derive(Serialize, Deserialize, Debug, Builder, Clone)]
 #[builder(field(public))]
